# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AprovaJá API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /study-plans:
    get:
      operationId: getStudyPlans
      tags: [study-plans]
      summary: List user study plans
      responses:
        "200":
          description: Study plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyPlan"
    post:
      operationId: createStudyPlan
      tags: [study-plans]
      summary: Create a new study plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyPlanInput"
      responses:
        "201":
          description: Created study plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPlan"
  /study-plans/{id}:
    get:
      operationId: getStudyPlan
      tags: [study-plans]
      summary: Get a specific study plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Study plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPlan"
    patch:
      operationId: updateStudyPlan
      tags: [study-plans]
      summary: Update a study plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyPlanUpdate"
      responses:
        "200":
          description: Updated study plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPlan"
    delete:
      operationId: deleteStudyPlan
      tags: [study-plans]
      summary: Delete a study plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /study-plans/{id}/sessions:
    get:
      operationId: getStudySessions
      tags: [study-plans]
      summary: Get study sessions for a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Study sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudySession"
    post:
      operationId: createStudySession
      tags: [study-plans]
      summary: Log a study session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudySessionInput"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySession"

  # ── Flashcards ─────────────────────────────────────────────────────────────
...
  /simulados:
    get:
      operationId: getSimulados
      tags: [simulados]
      summary: List available mock exams
      responses:
        "200":
          description: Mock exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Simulado"
    post:
      operationId: createSimulado
      tags: [simulados]
      summary: Create a mock exam attempt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimuladoInput"
      responses:
        "201":
          description: Created simulado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Simulado"
  /simulados/{id}:
    get:
      operationId: getSimulado
      tags: [simulados]
      summary: Get a simulado with questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Simulado detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimuladoDetail"
  /simulados/{id}/submit:
    post:
      operationId: submitSimulado
      tags: [simulados]
      summary: Submit simulado answers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimuladoAnswers"
      responses:
        "200":
          description: Simulado result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimuladoResult"
...
components:
  schemas:
...
    SubjectScore:
      type: object
      properties:
        subject:
          type: string
        score:
          type: number
        totalQuestions:
          type: integer
        correctAnswers:
          type: integer
        trend:
          type: string
          description: up, down, stable
      required: [subject, score, totalQuestions, correctAnswers, trend]
    StudyPlan:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        goal:
          type: string
        targetDate:
          type: ["string", "null"]
        subjects:
          type: array
          items:
            type: string
        hoursPerDay:
          type: number
        progress:
          type: number
        createdAt:
          type: string
      required: [id, title, goal, subjects, hoursPerDay, progress, createdAt]
    StudyPlanInput:
      type: object
      properties:
        title:
          type: string
        goal:
          type: string
        targetDate:
          type: ["string", "null"]
        subjects:
          type: array
          items:
            type: string
        hoursPerDay:
          type: number
      required: [title, goal, subjects, hoursPerDay]
    StudyPlanUpdate:
      type: object
      properties:
        title:
          type: string
        goal:
          type: string
        targetDate:
          type: ["string", "null"]
        subjects:
          type: array
          items:
            type: string
        hoursPerDay:
          type: number
    StudySession:
      type: object
      properties:
        id:
          type: integer
        planId:
          type: integer
        subject:
          type: string
        durationMinutes:
          type: integer
        notes:
          type: ["string", "null"]
        xpEarned:
          type: integer
        createdAt:
          type: string
      required: [id, planId, subject, durationMinutes, xpEarned, createdAt]
    StudySessionInput:
      type: object
      properties:
        subject:
          type: string
        durationMinutes:
          type: integer
        notes:
          type: ["string", "null"]
      required: [subject, durationMinutes]
...
    Simulado:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          description: ENEM, FUVEST, Concurso, Custom
        subject:
          type: ["string", "null"]
        questionCount:
          type: integer
        durationMinutes:
          type: integer
        difficulty:
          type: string
          description: facil, medio, dificil
        completedAt:
          type: ["string", "null"]
        score:
          type: ["number", "null"]
        createdAt:
          type: string
      required: [id, title, type, questionCount, durationMinutes, difficulty, createdAt]
    SimuladoInput:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
        subject:
          type: ["string", "null"]
        difficulty:
          type: string
      required: [title, type, difficulty]
    SimuladoDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
        difficulty:
          type: string
        durationMinutes:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
      required: [id, title, type, difficulty, durationMinutes, questions]
    Question:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        subject:
          type: string
        alternatives:
          type: array
          items:
            $ref: "#/components/schemas/Alternative"
      required: [id, text, subject, alternatives]
    Alternative:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
      required: [id, text]
    SimuladoAnswers:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerItem"
      required: [answers]
    AnswerItem:
      type: object
      properties:
        questionId:
          type: integer
        selectedAlternative:
          type: string
      required: [questionId, selectedAlternative]
    SimuladoResult:
      type: object
      properties:
        id:
          type: integer
        simuladoId:
          type: integer
        simuladoTitle:
          type: string
        score:
          type: number
        correctCount:
          type: integer
        totalCount:
          type: integer
        xpEarned:
          type: integer
        timeSpentMinutes:
          type: integer
        completedAt:
          type: string
        subjectBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/SubjectScore"
      required: [id, simuladoId, simuladoTitle, score, correctCount, totalCount, xpEarned, timeSpentMinutes, completedAt, subjectBreakdown]
...