# lib/api-zod/src/generated/types/simulado.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AprovaJá API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Simulado {
  id: number;
  title: string;
  /** ENEM, FUVEST, Concurso, Custom */
  type: string;
  /** @nullable */
  subject?: string | null;
  questionCount: number;
  durationMinutes: number;
  /** facil, medio, dificil */
  difficulty: string;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  score?: number | null;
  createdAt: string;
}