# lib/api-zod/src/generated/types/simuladoResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AprovaJá API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubjectScore } from './subjectScore';

export interface SimuladoResult {
  id: number;
  simuladoId: number;
  simuladoTitle: string;
  score: number;
  correctCount: number;
  totalCount: number;
  xpEarned: number;
  timeSpentMinutes: number;
  completedAt: string;
  subjectBreakdown: SubjectScore[];
}