# lib/api-zod/src/generated/types/studyPlan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AprovaJá API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudyPlan {
  id: number;
  title: string;
  goal: string;
  /** @nullable */
  targetDate?: string | null;
  subjects: string[];
  hoursPerDay: number;
  progress: number;
  createdAt: string;
}