# lib/api-zod/src/generated/types/studySession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AprovaJá API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudySession {
  id: number;
  planId: number;
  subject: string;
  durationMinutes: number;
  /** @nullable */
  notes?: string | null;
  xpEarned: number;
  createdAt: string;
}