# lib/api-zod/src/generated/types/subjectScore.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AprovaJá API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubjectScore {
  subject: string;
  score: number;
  totalQuestions: number;
  correctAnswers: number;
  /** up, down, stable */
  trend: string;
}